CM26T-Enhanced_services_and_controllers-SE171270

+++ InteriorCoffeeAPIs/Controllers/AccountController.cs
@@ -13,7 +13,6 @@ using System.Collections.Generic;
 using JsonSerializer = System.Text.Json.JsonSerializer;
 using System.Threading.Tasks;
 using System.Text.Json;
-using InteriorCoffee.Application.Utils;
 
 namespace InteriorCoffeeAPIs.Controllers
 {
@@ -30,32 +29,29 @@ namespace InteriorCoffeeAPIs.Controllers
         }
 
         [HttpGet(ApiEndPointConstant.Account.AccountsEndpoint)]
-        [ProducesResponseType(typeof(IPaginate<Account>), StatusCodes.Status200OK)]
-        [SwaggerOperation(Summary = "Get all accounts with pagination and sorting. " +
-            "Ex url: GET /api/accounts?pageNo=1&pageSize=10&sortBy=username&ascending=true\r\n")]
-        public async Task<IActionResult> GetAccounts([FromQuery] int? pageNo, [FromQuery] int? pageSize, [FromQuery] string sortBy = null, [FromQuery] bool? ascending = null)
+        [ProducesResponseType(typeof(AccountResponseDTO), StatusCodes.Status200OK)]
+        [SwaggerOperation(Summary = "Get all accounts with pagination, sorting, and filtering. " +
+            "Ex url: GET /api/accounts?pageNo=1&pageSize=10&sortBy=username&ascending=true&roleId=123&status=active&keyword=john\r\n")]
+        public async Task<IActionResult> GetAccounts([FromQuery] int? pageNo, [FromQuery] int? pageSize, [FromQuery] string sortBy = null, [FromQuery] bool? ascending = null,
+                                                     [FromQuery] string roleId = null, [FromQuery] string status = null, [FromQuery] string keyword = null)
         {
-            //Check filter options
             OrderBy orderBy = null;
             if (!string.IsNullOrEmpty(sortBy))
             {
                 orderBy = new OrderBy(sortBy, ascending ?? true);
             }
 
-            //Get data
-            var (accounts, currentPage, currentPageSize, totalItems, totalPages) = await _accountService.GetAccountsAsync(pageNo, pageSize, orderBy);
-
-            var response = new Paginate<Account>
+            var filter = new AccountFilterDTO
             {
-                Items = accounts,
-                PageNo = currentPage,
-                PageSize = currentPageSize,
-                TotalItems = totalItems,
-                TotalPages = totalPages,
+                Status = status,
+                RoleId = roleId
             };
+
+            var response = await _accountService.GetAccountsAsync(pageNo, pageSize, orderBy, filter, keyword);
 
             return Ok(response);
         }
+
 
 
         [HttpGet(ApiEndPointConstant.Account.AccountEndpoint)]
@@ -64,6 +60,10 @@ namespace InteriorCoffeeAPIs.Controllers
         public async Task<IActionResult> GetAccountById(string id)
         {
             var result = await _accountService.GetAccountByIdAsync(id);
+            if (result == null)
+            {
+                return NotFound();
+            }
             return Ok(result);
         }
 
@@ -92,6 +92,10 @@ namespace InteriorCoffeeAPIs.Controllers
         public async Task<IActionResult> UpdateAccount(string id, [FromBody] JsonElement updateAccount)
         {
             var existingAccount = await _accountService.GetAccountByIdAsync(id);
+            if (existingAccount == null)
+            {
+                return NotFound(new { Message = "Account not found" });
+            }
 
             var schemaFilePath = "AccountValidate"; // Use the correct key
             var validationService = _validationServices[schemaFilePath];
@@ -100,7 +104,7 @@ namespace InteriorCoffeeAPIs.Controllers
             var existingAccountJson = JsonSerializer.Serialize(existingAccount, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.KebabCaseLower });
             var existingAccountElement = JsonDocument.Parse(existingAccountJson).RootElement;
 
-            var mergedAccount = JsonUtil.MergeJsonElements(existingAccountElement, updateAccount);
+            var mergedAccount = MergeJsonElements(existingAccountElement, updateAccount);
 
             var jsonString = JsonSerializer.Serialize(mergedAccount, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.KebabCaseLower });
             var (isValid, errors) = validationService.ValidateJson(jsonString);
@@ -117,50 +121,50 @@ namespace InteriorCoffeeAPIs.Controllers
             return Ok("Action success");
         }
 
-        #region Clean Code (Use Util instead)
-        //private JsonElement MergeJsonElements(JsonElement original, JsonElement update)
-        //{
-        //    using (var stream = new MemoryStream())
-        //    {
-        //        using (var writer = new Utf8JsonWriter(stream))
-        //        {
-        //            writer.WriteStartObject();
+        private JsonElement MergeJsonElements(JsonElement original, JsonElement update)
+        {
+            using (var stream = new MemoryStream())
+            {
+                using (var writer = new Utf8JsonWriter(stream))
+                {
+                    writer.WriteStartObject();
 
-        //            foreach (var property in original.EnumerateObject())
-        //            {
-        //                if (update.TryGetProperty(property.Name, out var updatedProperty))
-        //                {
-        //                    writer.WritePropertyName(property.Name);
-        //                    updatedProperty.WriteTo(writer);
-        //                }
-        //                else
-        //                {
-        //                    property.WriteTo(writer);
-        //                }
-        //            }
+                    foreach (var property in original.EnumerateObject())
+                    {
+                        if (update.TryGetProperty(property.Name, out var updatedProperty))
+                        {
+                            writer.WritePropertyName(property.Name);
+                            updatedProperty.WriteTo(writer);
+                        }
+                        else
+                        {
+                            property.WriteTo(writer);
+                        }
+                    }
 
-        //            foreach (var property in update.EnumerateObject())
-        //            {
-        //                if (!original.TryGetProperty(property.Name, out _))
-        //                {
-        //                    writer.WritePropertyName(property.Name);
-        //                    property.Value.WriteTo(writer);
-        //                }
-        //            }
+                    foreach (var property in update.EnumerateObject())
+                    {
+                        if (!original.TryGetProperty(property.Name, out _))
+                        {
+                            writer.WritePropertyName(property.Name);
+                            property.Value.WriteTo(writer);
+                        }
+                    }
 
-        //            writer.WriteEndObject();
-        //        }
+                    writer.WriteEndObject();
+                }
 
-        //        stream.Position = 0;
-        //        using (var document = JsonDocument.Parse(stream))
-        //        {
-        //            return document.RootElement.Clone();
-        //        }
-        //    }
-        //}
-        #endregion
+                stream.Position = 0;
+                using (var document = JsonDocument.Parse(stream))
+                {
+                    return document.RootElement.Clone();
+                }
+            }
+        }
 
-        [HttpPatch(ApiEndPointConstant.Account.SoftDeleteAccountEndpoint)]
+
+
+        [HttpPut(ApiEndPointConstant.Account.SoftDeleteAccountEndpoint)]
         [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
         [SwaggerOperation(Summary = "Soft delete an account")]
         public async Task<IActionResult> SoftDeleteAccount(string id)
@@ -175,6 +179,10 @@ namespace InteriorCoffeeAPIs.Controllers
         public async Task<IActionResult> DeleteAccount(string id)
         {
             var account = await _accountService.GetAccountByIdAsync(id);
+            if (account == null)
+            {
+                return NotFound();
+            }
 
             await _accountService.DeleteAccountAsync(id);
             return Ok("Action success");
